--- src/main/java/duke/commands/Bye.java
@@ -1,10 +1,0 @@
-package duke.commands;
-
-import duke.data.MyData;
-import duke.ui.Ui;
-
-public class Bye extends Command {
-    public void execute(MyData data, Ui ui) {
-        ui.bye();
-    }
-}
