--- src/main/java/duke/commands/List.java
@@ -1,10 +1,0 @@
-package duke.commands;
-
-import duke.data.MyData;
-import duke.ui.Ui;
-
-public class List extends Command {
-    public void execute(MyData data, Ui ui) {
-        ui.list(data);
-    }
-}
