+++ src/main/java/pix/tasks/ToDo.java
@@ -1,0 +1,18 @@
+package duke.tasks;
+
+public class ToDo extends Task {
+    /** Constructs a new To Do object.*/
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String toSave() {
+        return "T /" + super.toSave();
+    }
+}
