+++ src/test/java/pix/ByeTest.java
